Add free list index
The free list index keeps track of unused addresses in the data segment of the database file. This removes the need to preallocate disk space as the database can now store data in arbitrary (or sequential) positions, regardless of the key.
Also, the primary indexes now store the address (or position) of the data within the file which should speed up data select.